API/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
pydantic==2.10.3
pydantic-core==2.27.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2